requirements.txt
langchain
langchain-openai
langchain-google-genai
langchain-community
langchain-core
python-dotenv
